Fix an issue with the cmake export files
bug 5518

Johannes Ring noted that setting the Trilinos_INSTALL_INCLUDE_DIR variable to
something other than the default would cause the Cmake export files to have the
wrong path to the header files. This was because of a change that had been made
to support binary installers such as the windows installer that can relocate
files at install time.

The solution was to append the value of Trilinos_INSTALL_INCLUDE_DIR to the
relative path that had been added in. This change however, does lock us into
only supporting relative paths for the set of variables
Trilinos_INSTALL_INCLUDE_DIR, Trilinos_INSTALL_LIB_DIR and
Trilinos_INSTALL_RUNTIME_DIR. Doing so does fix a few other issues that were
noted when investigating this and I believe that to be the primary use case for
those variables so we aren't really losing much. Handling non-relative paths
there would require significant changes to be able to support both relocatable
installers and the simpler installers like "make install".