STK_Mesh: Examine and unit-test induced-part concept
Confusion with induced-parts in STK_Mesh has, once again, caused problems
on the stk_migration branch. There are two issues:
* Transitivity of inducd-parts
* Superset/subset correctness with induced parts

Work on the migration branch exposed that induced parts were not transitive
in STK (IE if you had elem->edge->node, elem parts would be induced onto the
edge but not the node). I thought this was a mistake and changed STK to have
transitive induced parts, but it turns out that the domain document explicitly
says induced parts should not be transitive, so I have changed the behvior back.

The superset issue is even more hairy. As it turns out, the current induced-part
behavior in STK is not superset/subset consistent. If a node is induced into ElemPart,
it will not be added to the supersets of ElemPart, which is not consistent. This is
relatively easy to fix, but an ambiguity arises when relations are removed. If the relation
is severed, should node be removed from all supersets of ElemPart? Probably not, since
this would do strange things like remove it from the universal part. The problem
is that we have no way of knowing whether the node's membership in an unranked
superset of ElemPart was due solely to induced parts. The quick-and-dirty solution
is to only de-induced ranked parts. This puts us in the situation where add/remove
part will not be symmetric with respect to part membership. This is already the
case with add/remove part.

To help address these issues, I have written a set of unit tests that test the
induced-part mechansisms. I have also added superset/subset invariant checking
to BulkData.

I think we need to have a team discussion about this, but I have to push
forward with this now so I can merge this onto the migration branch since
this issue is breaking things there.

(From Sierra)