Update Kalli driver scripts
Kalli has been behaving pretty well recently so updating the scripts in
preparation of switching all nightly testing on kalli to these TDD scripts.

Added a section for release testing, though the release branch has not been
tested yet. This is more to get a starting point for the release testing ready
in preparation of testing it. It won't be enabled just yet.