testing harness: remove old cron scripts
A long overdue cleaning up of the old cron scripts used to run the various test
builds. Currently pretty much every test machine should be using the TDD to
drive its testing so these are no longer needed and it is confusing to have them
around.

There is one exception to this which is Kallikrates. It will be moving to TDD
soon for most of its testing, but several of the old style scripts are still in
use, and might need to be for a bit longer.