Update boost path on trilinos-test2
The boost that was being used was installed in the same place as the openmpi
compiler which was unfortunately allowing access to an actual mpi.h header to
packages that needed the siMPI mpi.h header. This was causing Isorropia and all
packages dependent on it to have unresolved symbols to MPI functions in serial
builds. The boost we are using now has previously been tested to work and more
importantly does not share an include directory with openmpi so won't
inadvertently cause real mpi.h headers to be pulled in for serial builds.