Added original matrix-free SG approach plus some other minor changes.
To better compare the product tensor approach I added the original matrix-free
SG multiply algorithm, which currently requires stokhos.  To handle this I
changed the build script to have a "stokhos" option, which requires passing
in the path to top-level of a Trilinos build enabling stokhos.  I also changed
the ranges of polynomial order and stochastic dimension, and disabled some
of the approaches in the testing output (it currently matches the data I used
for the Copper Mountain presentation).  To support this, I had to make a
couple of changes to the internals of Kokkos:

* Add support for a multivector update method, i.e., a saxpy
* Add support for crs matrix times multivector for host and cuda.
* Add support for writing out a crs matrix to a matrix market file (host only)

Hopefully this doesn't screw up any of Carter's testing/development.  If so,
we can always go back to the original TestStochastic.hpp file.  However I would
like to keep this one around so we can regenerate the data for the presentation
if necessary.