Adding reads for signed and unsigned long and short int types
I added code to read the various signed and unsigned short and long int types.
I was hoping to be able to reproduce a failure that we are seeing in Drekar
that reads in unsigned int parameters on GCC 4.6.1.  This particular build has
failed to produce the failures, at least in the MPI_DEBUG build that I tried.