Added logic to not do a global rebuild if cmake/ExtraRepositoriesList.cmake is modified
There is no reason to trigger a global rebuild of this file is modified.
Actually, there might be a reason with the CI server but it is not a great
reason.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework, VERAIn
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=96,notpassed=0 (2.81 min)
1) SERIAL_RELEASE => passed: passed=96,notpassed=0 (2.47 min)
2) MPI_DEBUG_SS => passed: passed=97,notpassed=0 (4.26 min)
3) SERIAL_RELEASE_SS => passed: passed=97,notpassed=0 (3.01 min)