STK: Fix some conditional code that wasn't properly protected
There was an initialization that should only be done when building with Sierra
that wasn't properly protected by a define.

Note: this change had already been made on the Sierra side, however, to be able
to build STK it had to be made on the Trilinos side too. This will likely cause
a conflict when that commit is reached. In that case the changes from Sierra
should take precedence.