Update brain's version of boost and add netcdf and hdf5
Updated boost to 1.49.0

added netcdf 4.2 and hdf5 1.8.6.

Note: These files were setup it looks like to be machine independent. However,
these changes will break that. The issue is that to be machine independent these
files were assuming that everything was available inside the hudson users home
dir. I didn't want to spend the time installing these tpls to wind up with them
only usable by our nightly testing so I put them in a more generic place. I
believe that this is okay at least in the short term as we have a few issues
that need to be worked out before platform independent builds are a reality.