SEACAS: update cmake files to reflect changes in Sierra
the decomp script was changed such that it no longer needs to be configured
before being usable so the configure step needed to be removed.

Two scripts, blot and fastq were added to the scripts directory so I added them
to the install call too.

Also updated the install call to a non-deprecated cmake function. This wasn't
causing issues, but if cmake ever drops compatibility for the deprecated call it
would be so I fixed it now instead of waiting to be bit by it.