Adding global option ${PROJECT_NAME}_INSTALL_LIBRARIES_AND_HEADERS
For default TriBITS projects, this defaults to TRUE but a TriBITS project and
default this to FASLE.  Since the CASL VERA TriBITS project is primarily
delivering executables and scripts, it does not make sense to install the
headers and libraries.  The only case where libraries will always be installed
is when BUILD_SHARED_LIBS=TRUE.

I have manually tested this several different ways.

We really need to write documentation and unit tests for the
TRIBITS_ADD_LIBRARY(...) function.  This will not be easy to do since it calls
a bunch of CMake functions that will need to be wrapped and tested somehow.
Technical debt ...