Piro: Made RythmosSolver more amenable to testing
Previously, the lone constructor of RythmosSolver required a deeply
nested parameter list in order to be able to instantiate the
internal objects required by the object under construction.

While such a design maximizes encapsulation, it nevertheless makes
testing more awkward than necessary. It also mixes up the primary
functionality of RythmosSolver with the responsibility of building its
own dependencies. This role is better delegated to a Factory-type
objects and/or functions.

The newly added constructor follows the dependency inversion principle
more closely.