Fixed Cijk tensors in pseudospectral expansion unit tests.
I accidentally copied the old way of setting the size of Cijk in
computeTripleProductTensor() in these tests, which was causing invalid
iterator dereferences (since the k-predicates were wrong).  Thanks to
Heidi for catching this with a checked-STL build.

I also added tests for this in the TotalOrder/TensorProduct/Smolyak
bases since this is an easy mistake to make until I port all of Stokhos
to this change.