Always validating the object index, even when not in debug mode (VRI Kanban #2423)
I have made the runtime check for an invalid object index non-optional to
check errors better and make it easier to write unit tests.  This is not an
expensive check and is likely small overhead compared to everything else.
Safe is better than fast for a class like this.

I made this change because I am writing unit tests for the Fortran PL wrappers
which use this class and they were failing in a non-debug build because they
expected the check of an invalid object to be always turned on (see VRI Kanban
story #2423).

Build/Test Cases Summary
Enabled Packages: TeuchosCore, CTeuchos, ForTeuchos
Disabled Packages: CTrilinos,ForTrilinos,PyTrilinos,Didasko,Mesquite,Phdmesh,Pliris,Claps
0) MPI_DEBUG => passed: passed=41,notpassed=0 (2.42 min)
1) SERIAL_RELEASE => passed: passed=41,notpassed=0 (1.20 min)
2) MPI_DEBUG_SS => passed: passed=83,notpassed=0 (2.57 min)
3) SERIAL_RELEASE_SS => passed: passed=81,notpassed=0 (1.67 min)