stk: alternative memory usage reporting
Add a memory reporting function for the current usage and the
high-water-mark.

This function includes both stack and heap allocations because it's
based on the resident set size (memory *actually* used by the
application). Different platforms have different system APIs and some
systems don't support native reporting of high-water-mark. Moreover,
if large amounts of stack memory are used in functions that go in and
out of scope, the reported memory can be non-intuitive (i.e., the
function may be out of scope but the memory is still reported as being
used). Fortunately, most of the memory usage in our applications is on
the heap and the reporting there seems more consistent with what you
see in code.

This also updates samba to use the new functions and moves the
handy human_bytes() function from samba into memory_util.hpp.

(From Sierra)