Adding basic support for APP-defined preconditioners to Thyra::ModelEvaluator
Here I have added support for a W_prec object of type
Thyra::PreconditionerBase to the Thyra::ModelEvaluator software.  There are
some simple tests that show that it works but more unit testing is needed (as
for the rest of the ME software stack).

There is the minimal support for W_prec in the other support software related
to Thyra::ModelEvaluator.

However, I did not add support for the pass-through of W_prec to the
EpetraExt::ModelEvaluator intrerface in the class Thyra::EpetraModelEvaluator.

Some of the software that I had to change is not currently tested in the Thyra
package but has some minimal tests in down-stream packages like MOOCHO.