Ifpack2: RILUK: Partial fix for Bug 5460.
RILUK already was storing and computing with the relevant parameters
as magnitude_type rather than double, but it was reading them from
the ParameterList as double.

Now setParameters first tries to read them from the ParameterList as
magnitude_type, and then (if that throws InvalidParameterType) tries to
read them as double (for backwards compatibility).