stk::mesh::impl  Partition ready for use; has some hooks for experimentation/tuning.
It is not clear that Partition::compress(.) should mimick the behavior of the
inner loop of the previous BucketRepository::optimize_buckets() implementation.
In particular, the heuristic of skipping single-bucket Partitions (bucket families,
in old parlance) has unintended consequences.  #defines in Partition.cpp can be
changed for some control of policy of when to sort and when and how far to compress
the resulting bucket's capacity.  These hooks and others have been useful for
investigating whether regression tests were sensitive to sorting or whether they
held pointers inappropriately.

(From Sierra)