Panzer: Added methods to setup provided DOFs in the default equation set
Previously there was a set of members that the user had to initialize.
This change is backward compatible and so the user can still modify
those methods, however, now there are a set of concrete functions to
call that support a broader range of capability. Most critically an
equation set have fields that are transient mixed with those that
are time independent. This is to support the use case like in Navier-Stokes
where the momentum equation and the div free constraint both appear
together.