Cleanup: Removed SIERRA- and Sandia-specific, moved push-log out
Here I did a few things to clean up these scripts while I was looking them
over to see how to extend them (to add white- and black-listing of
committers/pushers).

Here is what I did:

(*) Removed any referenes to SIERRA which is where these scripts were copied
from.

(*) Removed some specific references to software.sandia.gov.  There is still
some references to HTML documentation on software.sandia.gov so I left that
for now.

(*) Moved the push.log update script from post-recieve-email to post-recieve.
This is the right place to call this and shows how easy it is to add new
post-receive hooks.

I manually tested these changes locally.  If these scripts get any more
complicated we will need to create some unit tests.  This is not
self-sustaining software at this point (but it is also not all the large or
complex so this is not that big of a deal yet).