Kokkos,Tpetra::MultiVector: Specialized replace,sumInto for CPU Nodes.
Tpetra::MultiVector used to need a lot of function calls for
replaceLocalValue, sumIntoLocalValue, replaceGlobalValue, and
sumIntoGlobalValue.  This actually showed up in some recent performance
performance measurements by Matt Bettencourt for finite element
assembly using local indices.  Most of the overhead came from getting a
host view of a device pointer to the local multivector's data.  Also,
the code was messy and complicated.

I fixed the "messy and complicated" problem by adding replaceLocalValue
and sumIntoLocalValue methods to Kokkos::MultiVector.  Now the Tpetra
methods of the same name just defer to their Kokkos counterparts, and
the *GlobalValue methods in Tpetra just call Tpetra's corresponding
*LocalValue method (after converting the global row index to a local
row index).

I fixed the "lots of function calls problem" in part by the new Kokkos
methods, and in part by adding partial specializations of
Kokkos::MultiVector for specific CPU Node types.  The only methods
that are different in the partial specializations are replaceLocalValue
and sumIntoLocalValue.  Right now I only have a specialization for
Node=Kokkos::SerialNode, but this will change soon.  Adding
specializations for CPU Node types is just a matter of copy and paste.
The default implementation of these methods still does the right thing
for any Node type, include GPU Node types.  (It cleverly only copies
the one value that was changed or updated, and it uses a write-only
host buffer for replace.)