PARSER: Add optional parsing of input file via aprepro
The parsing of the input file with aprepro is enabled with the
"--aprepro" or "-a" command line option.  Aprepro variables can be
defined with the "--define" or "-D" option.  The syntax is:

 --define "var=value var2=value var3=string"  OR
 --define "var=value,var2=value,var3=string"

for example,

 adagio --aprepro  --define "strain_increment_method=midpoint_increment preconditioner=globalprobe" \
         --input input_file.i

The defined variables currently must be of the form "var=value" where
value is either an integer, a floating point number, or a string.  A
string value cannot have embedded spaces or an equals sign.  For
example, the following are invalid:

	 var="a value with spaces"  (INVALID)
	 var="test=new_value"  (INVALID)

The define syntax is very basic at this time and could be given more
capability if needed.

Note that this currently only affects the Sierra executables and is
not yet supported by the sierra script.  The script will be modified
to support this capability soon.

Several adagio regression tests have been modified to test this
capability.

(From Sierra)