Remove usage of RCP_STRENGTH_INVALID=-1 causing compile warning/error with GCC 4.7.2 (Trilinos bug #5824)
The issue here is that when you build Teuchos with GCC 4.7.2 with
-Werror=array-bounds, it compains that a statically sized array might be
incremented out of bounds by an enum for the value RCP_STRENGTH_INVALID=-1.

This was never causing a runtime memmory access error because this value was
used to detect invalid usage and was well unit tested.  The current version of
the code is slightly less secure from a testing standpoint but if the compiler
is happy I am fine with that.

Mike Glass manually tested this change and confirmed that it resolved the
issue.  See Trilinos bug #5824 for more details.